Refactor for typescript
