+++ .gitignore
@@ -55,3 +55,7 @@ buck-out/
 # Bundle artifact
 *.jsbundle
 ios/Pods/
+
+# Jest
+#
+.jest/
