+++ package.json
@@ -32,6 +32,7 @@
     "@types/jest": "^25.1.3",
     "@types/react": "^16.9.23",
     "@types/react-native": "^0.61.20",
+    "@types/react-native-modal-dropdown": "^0.6.1",
     "@types/react-test-renderer": "^16.9.2",
     "babel-jest": "^25.1.0",
     "babel-plugin-module-resolver": "^4.0.0",
@@ -40,9 +41,25 @@
     "metro-react-native-babel-preset": "^0.58.0",
     "react-native-typescript-transformer": "^1.2.13",
     "react-test-renderer": "16.9.0",
+    "ts-jest": "^25.2.1",
     "typescript": "^3.8.3"
   },
   "jest": {
-    "preset": "react-native"
+    "preset": "react-native",
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js"
+    ],
+    "transform": {
+      "^.+\\.(js)$": "<rootDir>/node_modules/babel-jest",
+      "\\.(ts|tsx)$": "<rootDir>/node_modules/ts-jest/preprocessor.js"
+    },
+    "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
+    "testPathIgnorePatterns": [
+      "\\.snap$",
+      "<rootDir>/node_modules/"
+    ],
+    "cacheDirectory": ".jest/cache"
   }
 }
