--- src/common/components/buttons/Close.js
@@ -1,41 +1,0 @@
-import React, { Component } from 'react';
-import { Image, StyleSheet, TouchableOpacity } from 'react-native';
-import colors from '../../../common/theme/colors';
-import R from '@R/R';
-
-class Close extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
-  render() {
-    const { navigation } = this.props;
-
-    return (
-      <TouchableOpacity
-        style={styles.button}
-        onPress={() => navigation.goBack()}
-      >
-        <Image
-          style={styles.image}
-          resizeMode="contain"
-          source={R.images.ic_close}
-        />
-      </TouchableOpacity>
-    );
-  }
-}
-
-export default Close;
-
-const styles = StyleSheet.create({
-  button: {
-    justifyContent: 'center',
-    alignItems: 'center'
-  },
-  image: {
-    marginLeft: 10,
-    tintColor: colors.onBackground
-  }
-});
