+++ src/resources/assets/common/icons.ts
