+++ src/resources/colors/index.ts
