+++ tsconfig.json
@@ -43,6 +43,7 @@
     "baseUrl": "./",                       /* Base directory to resolve non-absolute module names. */
     "paths": {
       "*": ["src/*"],
+      "@R/*": ["src/resources/*"],
       "tests": ["tests/*"]
     },                           /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
